backend/disasters_management/requirements.txt
astroid==2.0.4
dataclasses==0.6
Django==2.1.1
django-sendgrid-v5==0.6.893
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djongo==1.2.30
future==0.16.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
PyJWT==1.6.4
pylint==2.1.1
pymongo==3.7.1
python-http-client==3.1.0
pytz==2018.5
sendgrid==5.6.0
six==1.11.0
sqlparse==0.2.4
typed-ast==1.1.0
wrapt==1.10.11
